+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1X5WV14J3BBQT514P824WZ"
+        "01KN3FZ0CM0996JSRH1Y92T6ZR"
     )
     r"""The unique identifier of the data source"""