+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN1X5WV0HBBCYTV0T67BJPN0"
+        "01KN3FZ0CKRYMM6G274CRVVF2Y"
     )
     r"""The unique identifier of the data source"""